Add command foreach, to call arbitraty cmd in every external dir
This works exactly as git submodule foreach, and this allows to call, for
example `git externals foreach git checkout`.

In case of error:
`git externals foreach` stops execution and fails at the first failed (!= 0)
exit code of the nested calls. In this case the exit code of  `git externals`
itself is set to the exit code of the failed command