Comment out git_svn_update_checkout
Performing git-svn-rebase will ALWAYS retrieve the last reivision of
the external. But that's not what we want when the externals is frozen.

Because git_svn_update_checkout is called for every 'update', the
initial one and the further ones, just commenting it out is only
a temporary solution but at least it allows us to work with frozen
git-svn externals on CI and local.